Fix bug 1598 by calling into the OOo C++ code to get the position of the cursor and convert pointers passed to and from Java from int to long

+++ vcl/java/com/sun/star/vcl/VCLEventQueue.java
@@ -380,7 +380,7 @@ public final class VCLEventQueue implements Runnable {
 	 *
      * @param frame the frame pointer
 	 */
-	void removeCachedEvents(int frame) {
+	void removeCachedEvents(long frame) {
 
 		synchronized (queueList) {
 			for (int i = 0; i < queueList.length; i++) {
